+++ internal/model/event_model.go
@@ -41,7 +41,7 @@ type EventPrice struct {
 	gorm.Model
 	EventID uint   `gorm:"not null"`
 	Name    string `gorm:"not null"` // e.g., "Presale", "VIP"
-	Price   int    `gorm:"not null"`
+	Price   int64  `gorm:"not null"`
 	Quota   int    `gorm:"not null"`
 	Tickets []Ticket
 }
